web/I_am_not_a_robot/requirements.txt
requests
fastapi==0.75.1
Pillow==9.1.0
PyJWT==2.3.0
python-multipart
sqlalchemy
